Use paragraphFn instead of <p> for 'text'
Summary:
For markedown blocks that aren't followed by a header/blockquote/etc,
they are rendered as 'text' instead of 'paragraph', which was rendering
a <p> tag even when a special paragraphFn renderer was specified. This
fixes that, so that clients can avoid rendering any <p> tags (which is
especially important because you cannot put <div>s in <p> tags, and
attempting to do so confuses the browser and react.

Test Plan:
Cross fingers